refactor: move continue-watching stub data to a separate file
- Extracted the `CONTINUE_WATCHING_STUB` and `ContinueWatchingItem` type into a new `continueWatchingStub.ts` file for better organization and maintainability.
- Updated `ContinueWatchingSection.tsx` to import the stub data from the new file.
- Cleaned up the `index.ts` file to reflect the new structure.
