+++ apps/frontend/src/components/glass/ContinueWatchingSection.tsx
@@ -3,47 +3,11 @@ import { Heart, PauseCircle } from "lucide-react";
 
 import { navLinkOutlineClass } from "@/lib/glass-ui";
 import { cn, getMovieImageUrl } from "@/lib/utils";
+import { CONTINUE_WATCHING_STUB, type ContinueWatchingItem } from "./continueWatchingStub";
 import { GlassPanel } from "./GlassPanel";
 import { PosterCard } from "./PosterCard";
 import { PrimaryButton } from "./PrimaryButton";
 import { ProgressBar } from "./ProgressBar";
-
-export type ContinueWatchingItem = {
-  id: number;
-  title: string;
-  posterPath: string;
-  progress: number;
-  durationLabel: string;
-  remainingLabel: string;
-};
-
-// UI stub (#315): mocked continue-watching data until playback progress is available from API.
-export const CONTINUE_WATCHING_STUB: ContinueWatchingItem[] = [
-  {
-    id: 550,
-    title: "Fight Club",
-    posterPath: "/pB8BM7pdSp6B6Ih7QZ4DrQ3PmJK.jpg",
-    progress: 42,
-    durationLabel: "2h 19m",
-    remainingLabel: "1h 20m restantes",
-  },
-  {
-    id: 278,
-    title: "The Shawshank Redemption",
-    posterPath: "/9cqNxx0GxF0bflZmeSMuL5tnGzr.jpg",
-    progress: 68,
-    durationLabel: "2h 22m",
-    remainingLabel: "45m restantes",
-  },
-  {
-    id: 680,
-    title: "Pulp Fiction",
-    posterPath: "/d5iIlFn5s0ImszYzBPb8JPIfbXD.jpg",
-    progress: 23,
-    durationLabel: "2h 34m",
-    remainingLabel: "1h 58m restantes",
-  },
-];
 
 type ContinueWatchingSectionProps = {
   items?: ContinueWatchingItem[];
@@ -109,7 +73,6 @@ export function MiniPlayerTile({ items = CONTINUE_WATCHING_STUB }: ContinueWatch
         <h2 className="text-lg font-semibold text-ink">Mini-player</h2>
       </div>
       <p className="text-xs text-ink-muted">
-         (sans player réel).
       </p>
       <div className="rounded-2xl border border-[var(--glass-border)] bg-[var(--glass-bg)] p-3">
         <div className="flex items-start gap-3">
